This logic adds wireless tether for root users to the biffmod build: http://code.google.com/p/android-wifi-tether/
Since the user may update or other operation in the future we just download
the binary from their website and place it in the update.zip as is.

* download will happen on running environment setup if the file is not found
* Install happens as the last step of 'make biffmod-ota'

The licence of the binary/code of android-wifi-tether is GPLv3 thus we ought to
not have any complaints of including it.

BiffMod users who find the app useful should be encouraged to donate
to the android-wifi-tether creators (as indicated by the app)